+++ junit/kubernetes-server-mock/src/test/java/io/fabric8/kubernetes/client/server/mock/ErrorStreamMessageTest.java
@@ -16,23 +16,18 @@
 
 package io.fabric8.kubernetes.client.server.mock;
 
-import io.fabric8.mockwebserver.internal.WebSocketMessage;
+import org.junit.jupiter.api.Test;
 
-import java.nio.charset.StandardCharsets;
+import static org.assertj.core.api.Assertions.assertThat;
 
-public class OutputStreamMessage extends WebSocketMessage {
+class ErrorStreamMessageTest {
 
-  static final byte OUT_STREAM_ID = 1;
-
-  public OutputStreamMessage(String body) {
-    super(0L, getBodyBytes(OUT_STREAM_ID, body), true, true);
-  }
-
-  static byte[] getBodyBytes(byte prefix, String body) {
-    byte[] original = body.getBytes(StandardCharsets.UTF_8);
-    byte[] prefixed = new byte[original.length + 1];
-    prefixed[0] = prefix;
-    System.arraycopy(original, 0, prefixed, 1, original.length);
-    return prefixed;
+  @Test
+  void testMessageEncoding() {
+    final ErrorStreamMessage message = new ErrorStreamMessage("foobar");
+    assertThat(message.isBinary()).isTrue();
+    assertThat(message.isToBeRemoved()).isTrue();
+    assertThat(message.getBytes()).startsWith(ErrorStreamMessage.ERR_STREAM_ID);
+    assertThat(message.getBody().substring(1)).isEqualTo("foobar");
   }
 }